Added Enumeration to the project

+++ EnumerationExercise2/EnumerationExercise2.java
@@ -49,7 +49,7 @@ class Shop{
     }
 
     public double discountCalculator(double amount, int discount){
-        return (amount - (amount * (float)discount/100));
+        return (amount - (amount * discount/100));
     }
 }
